requirements.txt
Django==2.2.*
djangorestframework==3.9.*
flake8==3.6.*